requirements.txt
Django
dlib
face_recognition
opencv-python
gunicorn
psycopg2-binary  # PostgreSQL driver
python-decouple
